fix: シェアードからインテント完了

+++ app/src/main/java/app/makino/harutiro/resito2/input/inputPageNedan.kt
@@ -78,7 +78,7 @@ class inputPageNedan : AppCompatActivity() {
 
             //インテント
             val inputPage = Intent(this,testInput::class.java)
-            inputPage.putExtra("nedanItiziDate",totalNumber)
+            inputPage.putExtra("nedanItiziDate",totalNumber.toInt())
             startActivity(inputPage)
             finish()
